+++ src/api/controllers/users.controller.js
@@ -1,6 +1,6 @@
 const { v4: uuidv4 } = require('uuid')
 
-const User = require('../../models/users.model')
+const User = require('../models/users.model')
 
 async function addUser (req, res, next) {
   try {
