+++ src/api/routes/users.router.js
@@ -2,7 +2,7 @@ const express = require('express')
 const router = express.Router()
 const {
   addUser
-} = require('./users.controller')
+} = require('../controllers/users.controller')
 
 // ADD USER
 router.post('/create', addUser)
